requirements.txt
tqdm>=4.46.0
crepe>=0.0.11
librosa>=0.8.0
matplotlib>=3.3.3
numpy>=1.19.4
PyYAML>=5.3.1
soundfile>=0.10.3.post1
torch>=1.7.0
tensorboard
tensorflow